Removing beforeunload EventListener on detach
This enables a View setup where the UnloadObserver is added to a child view "A". Once the user navigates to another child view "B" of the same MainView, the UnloadObserver must be gone. I could do this manually in child view "A" (and all other views with that behaviour) or the UnloadObserver could take care of this directly so I don't have to remember to add a detach listener to all my views that use an unloadObserver. I believe this does not break the component for those that attach the UnloadObserver to the UI directly.